+++ psakeFile.ps1
@@ -11,6 +11,8 @@ properties {
 
 task default -depends Test
 
+task Build -FromModule PowerShellBuild -Version '0.3.1'
+
 task Test -FromModule PowerShellBuild -Version '0.3.1'
 
 task PublishToPSGallery -Action {