+++ CHANGELOG.md
@@ -7,3 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.1.0] Unreleased
 
+* Created Invoke-RestMethod wrapper to include token upon each invoke-restmethod.
+* Added Get-CSToken to retrive token for OAuth2 authorization.
+* Added initial support for retrieving host details using Get-CSDeviceDetails.
+* Updated readme.md with basic instructions.